use per-method file structure for remote service.
